Moving Javascript Functions
