+++ web/tools.html
@@ -4,7 +4,7 @@
         <title>Tools | Rerum Playground</title>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <script src="./js/tool.js" type="module" defer></script>        
+        <script src="./js/tools.js" type="module" defer></script>        
         <script src="./js/playground.js" type="module"></script>
         <link rel="stylesheet" href="styles.css">
         <link rel="stylesheet" href="https://unpkg.com/chota@latest">
